refactor: editModeSlice refactoring
