+++ src/components/member/index.ts
@@ -1,0 +1,1 @@
+export { default as MemberProfile } from './MemberProfile'
