+++ src/lib/api/Member/MemberAPI.ts
@@ -1,7 +1,6 @@
 import decodeJWT from 'src/lib/util/decodeJWT'
 import { EDIT_MEMBER_URL, USER_URL } from 'src/lib/constants/Url'
-import { type } from 'src/pages/memberInfo'
-import { AxiosRequestConfig } from 'axios'
+import { EditMemberParam } from 'src/lib/types/UserInterface'
 import HttpClient from '../httpClient'
 
 class MemberAPI extends HttpClient {
@@ -12,14 +11,7 @@ class MemberAPI extends HttpClient {
   }
 
   private _initializeResponseInterceptor = () => {
-    this.instance.interceptors.response.use(this._handleRequest)
-  }
-
-  protected _handleRequest = (config: AxiosRequestConfig | any) => {
-    if (config.headers) {
-      config.headers['Content-Type'] = 'multipart/form-data'
-    }
-    return config
+    this.instance.interceptors.response.use()
   }
 
   public getMemberInfo = async () => {
@@ -27,7 +19,7 @@ class MemberAPI extends HttpClient {
     return response.data
   }
 
-  public postMemberInfo = async (data: type) => {
+  public postMemberInfo = async (data: EditMemberParam) => {
     const profileUrl = data.profileUrl ? data.profileUrl : ''
     const payload = { imageFile: data.imageFile }
     const response = await this.instance.post(
